@@ -82,8 +82,8 @@
     "electron-squirrel-startup": "1.0.0",
     "file-saver": "1.3.8",
     "js-kinesis-sdk": "0.9.7",
-    "react": "16.3.2",
-    "react-dom": "16.3.2",
+    "react": "^16.4.1",
+    "react-dom": "^16.4.1",
     "react-hot-loader": "3.0.0-beta.6",
     "react-redux": "5.0.7",
     "redux": "4.0.0",
@@ -104,8 +104,8 @@
     "@types/electron-json-storage": "4.0.0",
     "@types/file-saver": "1.3.0",
     "@types/lodash": "4.14.108",
-    "@types/react": "16.3.13",
-    "@types/react-dom": "16.0.5",
+    "@types/react": "^16.4.6",
+    "@types/react-dom": "^16.0.6",
     "@types/react-redux": "6.0.0",
     "@types/redux-persist": "4.3.1",
     "electron": "1.8.4",
@@ -114,8 +114,10 @@
     "node-sass": "4.9.0",
     "parcel-bundler": "1.8.1",
     "parcel-plugin-typescript": "0.7.1",
+    "prettier": "^1.13.7",
     "tslint": "5.10.0",
+    "tslint-config-prettier": "^1.14.0",
     "tslint-react": "3.6.0",
-    "typescript": "2.8.3"
+    "typescript": "^2.9.2"
   }
 }
