@@ -1,18 +1,23 @@
-import { updateSignTransactionForm } from '@actions'
+import { transactionRequest, updateSignTransactionForm } from '@actions'
 import { SignTransactionForm as SignTransactionFormPresentation } from '@components'
 import { RootState } from '@store'
 import { Wallet } from '@types'
 import { connect } from 'react-redux'
 
-const mapStateToProps = ({ sign, passwords, wallets, connections }: RootState) => ({
+const mapStateToProps = ({ sign, passwords, wallets }: RootState) => ({
   message: sign.signTransactionData.message,
-  decryptedPrivateKey: () => passwords.livePasswords[(wallets.activeWallet as Wallet).publicKey].privateKey,
-  connection: connections.currentConnection,
+  decryptedPrivateKey: () =>
+    passwords.livePasswords[(wallets.activeWallet as Wallet).publicKey].privateKey,
 })
 
 const mapDispatchToProps = {
   updateSignTransactionForm,
+  transactionRequest,
 }
 
-export type SignTransactionFormProps = ReturnType<typeof mapStateToProps> & typeof mapDispatchToProps
-export const SignTransactionForm = connect(mapStateToProps, mapDispatchToProps)(SignTransactionFormPresentation)
+export type SignTransactionFormProps = ReturnType<typeof mapStateToProps> &
+  typeof mapDispatchToProps
+export const SignTransactionForm = connect(
+  mapStateToProps,
+  mapDispatchToProps,
+)(SignTransactionFormPresentation)
